Peter branch (#119)
* feat: init commit

* ngrok, local development, tests,market, quartz

* market fe, market api fix, mongo init script

* uncomment frontend docker compose

* economy fe + be

* economy ui fixes + matchreport ui + challenger UI + playertraining review

* enable FE in docker compose

+++ services/main-service/player-domain/src/main/java/com/kjeldsen/player/domain/generator/PointsGenerator.java
@@ -65,6 +65,6 @@ public class PointsGenerator {
 
     public static int generatePotentialPoints(int actualPoints) {
         Integer result = (int) (increaseMaxPercentage * actualPoints);
-        return (int) (Math.min(MAX_SKILL_VALUE, result + actualPoints));
+        return Math.min(MAX_SKILL_VALUE, result + actualPoints);
     }
 }
